pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rocketlaunch_feishu"
version = "0.1.0"
description = "A CLI tool for synchronizing rocket launch data to Feishu Bitable"
readme = "README.md"
requires-python = "<3.13" # 注意，您之前的 Python 版本是 >3.7 且 <3.12，这里是 <3.13
license = "MIT"
keywords = []
authors = [
  { name = "zhangchao", email = "zhangchao564265135@hotmail.com" },
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7", # 如果您确实支持3.7
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12", # 添加 3.12
]
dependencies = [
  "typer[all]>=0.9.0", # 使用 [all] 通常会包含 shell completion 等
  "rich>=13.0.0",
  "python-dotenv>=1.0.0", # 版本更新到 >=1.0.0 比较常见
  "beautifulsoup4>=4.12.0", # 版本更新
  "lark-oapi>=1.2.0", # 版本更新
  "pytest>=7.0.0", # 确保 pytest 是开发依赖或测试依赖
  "httpx>=0.25.0", # 您代码中用到了 httpx，但没在依赖中
  "zoneinfo; python_version < '3.9'", # 条件依赖 zoneinfo for Python < 3.9
]

# build = [  # 'build' 字段在 [project]下通常用于指定构建时的依赖，pyinstaller 更像是打包工具
#   "pyinstaller>=6.3.0",
# ]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    # "pyinstaller>=6.3.0", # 如果 PyInstaller 是开发/构建工具
]

[project.scripts]
rocketlaunch-feishu = "rocketlaunch_feishu.main:app" # 脚本名可以更具描述性

[tool.pytest.ini_options]
addopts = "-ra -q --color=yes" # 添加 --color=yes
testpaths = ["tests"]
python_files = ["test_*.py"]
pythonpath = ["src"]  # <--- 关键改动在这里